Update @foxglove/studio to v0.11.0
